tests pass with decapodes#migrate-diagrammatic-equations
