+++ src/DiagrammaticEquations.jl
@@ -3,26 +3,20 @@
 module DiagrammaticEquations
 
 export 
-DerivOp, append_dot, normalize_unicode, infer_states,
+DerivOp, append_dot, normalize_unicode, infer_states, infer_types!,
 # Deca
-op1_res_rules_1D,
-op2_res_rules_1D,
-op1_res_rules_2D,
-op2_res_rules_2D,
-op1_inf_rules_1D,
-op2_inf_rules_1D,
-op1_inf_rules_2D,
-op2_inf_rules_2D,
-recursive_delete_parents, findname, spacename, varname, unicode!, apply_inference_rule_op1!, apply_inference_rule_op2!,
+op1_res_rules_1D, op2_res_rules_1D, op1_res_rules_2D, op2_res_rules_2D, 
+op1_inf_rules_1D, op2_inf_rules_1D, op1_inf_rules_2D, op2_inf_rules_2D,
+recursive_delete_parents, findname, spacename, varname, unicode!, vec_to_dec!, 
 ## collages
 Collage, collate,
 ## composition
 oapply, unique_by, unique_by!, OpenSummationDecapodeOb, OpenSummationDecapode, Open,
 ## acset
 SchDecapode, SchNamedDecapode, AbstractDecapode, AbstractNamedDecapode, NamedDecapode, SummationDecapode,
-contract_operators, add_constant!, add_parameter, vec_to_dec!, infer_types!,
-fill_names!,  dot_rename!, expand_operators, infer_state_names, recognize_types,
+contract_operators, add_constant!, add_parameter, fill_names!,  dot_rename!, expand_operators, infer_state_names, recognize_types,
 resolve_overloads!, replace_names!,
+apply_inference_rule_op1!, apply_inference_rule_op2!, 
 ## language
 @decapode, Term, parse_decapode, term, Eq, DecaExpr,
 # ~~~~~
