+++ src/deca/Deca.jl
@@ -1,19 +1,15 @@
 module Deca
+
+using DataStructures
 using ..DiagrammaticEquations
 using Catlab
 
+import ..infer_types!, ..resolve_overloads!
+
+export normalize_unicode, varname, infer_states, infer_types!, resolve_overloads!, typename, spacename, recursive_delete_parents, recursive_delete_parents!, unicode!, op1_res_rules_1D, op2_res_rules_1D, op1_res_rules_2D, op2_res_rules_2D, op1_inf_rules_1D, op2_inf_rules_1D, op1_inf_rules_2D, op2_inf_rules_2D, vec_to_dec!
+
 include("deca_acset.jl")
 include("deca_visualization.jl")
-
-export normalize_unicode, varname, infer_states, infer_types!, typename, spacename, recursive_delete_parents, recursive_delete_parents!, unicode!, replace_names!, op1_res_rules_1D,
-op2_res_rules_1D,
-op1_res_rules_2D,
-op2_res_rules_2D,
-op1_inf_rules_1D,
-op2_inf_rules_1D,
-op1_inf_rules_2D,
-op2_inf_rules_2D,
-apply_inference_rule_op1!, apply_inference_rule_op2!
 
 ## TODO: where?
 function infer_states(d::SummationDecapode)
